Backend_API_Flask/requirements.txt
Flask
flask-cors
requests
google-auth
google_auth_oauthlib
google-api-python-client
feedparser
python-dotenv
